[Refactor] Move To Navi Provider
