--- app/src/main/res/values/dimens.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <!-- Font Size-->
-
-</resources>
