+++ navigation/src/main/java/com/angdroid/navigation/EditMumentNavigatorProvider.kt
@@ -1,4 +1,4 @@
-package com.mument_android.app.presentation.ui.detail.mument.navigator
+package com.angdroid.navigation
 
 import com.mument_android.domain.entity.detail.MumentDetailEntity
